+++ CONTRIBUTING.rst
