+++ day_summary.py
@@ -1,14 +1,15 @@
-#!/usr/bin/python
+ #!/usr/bin/env python
 
 import sqlite3
 import datetime
 
 def createTable(conn):
-	conn.execute('''CREATE TABLE if not exists SUMMARY_ATTENDANCE
-		(ID INTEGER PRIMARY KEY AUTOINCREMENT,
-		ROLL_NUMBER INTEGER NOT NULL,
-		DT datetime default (datetime(current_timestamp,'LOCALTIME')),
-		DURATION_HOURS FLOAT);''')
+	conn_cursor=conn.cursor()
+	conn_cursor.execute('''CREATE TABLE IF NOT EXISTS SUMMARY_ATTENDANCE
+		(id INTEGER PRIMARY KEY AUTOINCREMENT,
+		roll_number INTEGER NOT NULL,
+		dt datetime default (datetime(current_timestamp,'LOCALTIME')),
+		duration_hours FLOAT);''')
 
 def displayDB():
 	try:
@@ -16,28 +17,27 @@ def displayDB():
 	except:
 		conn=sqlite3.connect("test.db")
 	finally:
-		cursor = conn.execute("SELECT ID,ROLL_NUMBER,DT,DURATION_HOURS from SUMMARY_ATTENDANCE")
+		conn_cursor=conn.cursor()
+		cursor = conn_cursor.execute("SELECT id,roll_number,dt,duration_hours from SUMMARY_ATTENDANCE")
 		for row in cursor:
-	   		print "ID=", row[0],
-	   		print " ROLL_NUMBER=", row[1],
-   			print " DT=", row[2],
-   			print " DURATION_HOURS=",row[3]
+			print 'Id={} RollNumber={} Date/Time={} DurationHours={}'.format(row[0],row[1],row[2],row[3]);
    			
 # NOTE IT IS ASSUMED THAT EVERYONE HAS MARKED EXIT, EXACTLY BEFORE 2:00AM.
 # NOT MARKING JUST BEFORE 2:00 AM WOULD MARK HIM AS DEFAULTER.
-# EVERY ENTRY AFTER 2:00 AM WOULD BE AN "INCOMING ENTRY" ONLY
+# EVERY FIRST ENTRY AFTER 2:00 AM WOULD BE AN "INCOMING ENTRY" ONLY
 
 def getSummary():
-	Summary('NYO',8);
-	Summary('RTE',2);
+	Summary('nyo',8);
+	Summary('rte',2);
 
 def Summary(filename,hours):
 	conn=sqlite3.connect("test.db")
+	conn_cursor=conn.cursor()
 	list_students=open(filename,'r').readlines()
 	for student in list_students:
 		student=student.strip()
 		prevdate=datetime.datetime.now()-datetime.timedelta(days=1)
-		cursor=conn.execute("SELECT ID,ROLL_NUMBER,DT from ATTENDANCE where ROLL_NUMBER=? AND DT>?;",(int(student),str(prevdate)[:10]+" 02:00:00"))
+		cursor=conn_cursor.execute("SELECT id,roll_number,dt from ATTENDANCE where roll_number=? AND dt>?;",(int(student),str(prevdate)[:10]+" 02:00:00"))
 		#print str(prevdate)[:10]+" 02:00:00"
 		i=0;prev=None;
 		summ=0.0
@@ -56,13 +56,13 @@ def Summary(filename,hours):
 					summ=summ+duration.seconds/3600.0
 			i=i+1
 		createTable(conn)
-		conn.execute("INSERT INTO SUMMARY_ATTENDANCE (ROLL_NUMBER,DURATION_HOURS) \
-				VALUES (?,?)",(student,float("{0:.2f}".format(round(summ,2)))))
+		conn_cursor.execute("INSERT INTO SUMMARY_ATTENDANCE (roll_number,duration_hours) VALUES (?,?)",(student,float("{0:.2f}".format(round(summ,2)))))
 		if summ < hours :
 			# MAKE A DEFAULTER API CALL HERE
-			print student + " had less than "+str(hours)+" hours";
+			print '{} had less than {} hours'.format(student,str(hours));
 		if i%2==1:
-			print student+" didn't marked exit. Evaluated till last exit."
+			print "{} didn't marked exit. Evaluated till last exit.".format(student);
+			
 	return
