+++ docs/.gitignore
