+++ docs/authors.rst
