+++ docs/installation.rst
