+++ docs/readme.rst
